Add files via upload
Baseball Case Study

This project uses data from the 2014 Major League Baseball season to predict the number of wins for a team in the 2015 season. By analyzing various performance metrics, the aim is to identify key factors that contribute to a team's success and develop a model to forecast future wins.

Project Overview

The dataset includes 17 features, such as Runs Scored, Home Runs, Earned Run Average, and Errors, among others. These features provide insights into different aspects of a baseball team's performance. The target variable is the number of wins a team achieves in a season.

Key Steps in the Project:

1. Data Cleaning and Preprocessing:
   - Renamed columns for better clarity.
   - Detected and removed outliers to ensure accurate model performance.
   - Addressed skewness in the data using logarithmic transformation.

2.  Exploratory Data Analysis (EDA):
   - Visualized correlations between features and the target variable.
   - Used heatmaps and boxplots to understand data distribution and detect outliers.

3. Feature Selection and Dimensionality Reduction:
   - Analyzed feature correlations to handle multicollinearity using Variance Inflation Factor (VIF) and Principal Component Analysis (PCA).

4. Model Development:
   - Trained and evaluated multiple models including Linear Regression, Logistic Regression, and Decision Tree Classifier.
   - Performed Hyperparameter Tuning using GridSearchCV to optimize model performance.

5.  Model Evaluation:
   - The final model was a Linear Regression model with tuned hyperparameters, achieving a strong R2 score.
   - Evaluated the model using metrics like Mean Absolute Error (MAE), Mean Squared Error (MSE), and Root Mean Squared Error (RMSE).

6. Model Deployment:
   - Saved the final Linear Regression model using Joblib for future predictions.

 Conclusion

This project provided a comprehensive analysis of baseball team performance metrics and successfully predicted the number of wins for the 2015 season using a carefully tuned Linear Regression model.